requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
darkdetect==0.8.0
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
Django==4.2.7
django-cors-headers==4.3.1
django-storages==1.14.3
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.14.5
et-xmlfile==1.1.0
fonttools==4.49.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpie==3.2.2
httplib2==0.22.0
httpx==0.27.0
idna==3.4
img2pdf==0.5.1
isodate==0.6.1
janus==1.0.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msal==1.25.0
msal-extensions==1.0.0
multidict==6.0.5
mypy==1.7.1
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
paho-mqtt==1.6.1
pikepdf==8.11.2
Pillow==10.1.0
portalocker==2.8.2
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.35.1
requests>=2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requests-unixsocket==0.3.0
rich==13.7.0
rpds-py==0.18.1
rsa==4.9
seaborn==0.13.2
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
stripe==9.10.0
tabulate==0.9.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3>=1.26.5,<1.27
watchdog==4.0.0
wrapt==1.16.0
yarl==1.9.4
